+++ CHANGELOG.md
@@ -4,7 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
----
+## 2019/01/02
+### Changed
+- [scritps](/scripts)
+  - [config](/scripts/config)
+  - [macos.sh](/scripts/macos.sh)
+  - [macos-basic.sh](/scripts/macos-basic.sh)
+  - [macos-network.sh](/scripts/macos-network.sh)
+  - [macos-virsh.sh](/scripts/macos-virsh.sh)
+  - [windows.sh](/scripts/windows.sh)
+  - [windows-basic.sh](/scripts/windows-basic.sh)
+  - [windows-network.sh](/scripts/windows-network.sh)
+  - [windows-virsh.sh](/scripts/windows-virsh.sh)
+- [README.md](/README.md)
+
 
 ## 2018/12/30
 ### Added
@@ -45,7 +58,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - qemu@.service
   - test-qemu.sh
   - windows-install.sh
-- Troubleshoot.md [Moved to wiki]()
-- Install.md [Moved to wiki]()
-
----
+- Troubleshoot.md [Moved to wiki](https://gitlab.com/YuriAlek/vfio/wikis/Troubleshoot)
+- Install.md [Moved to wiki](https://gitlab.com/YuriAlek/vfio/wikis/Install)